# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Demo Swagger implementation for Book Management",
    "version": "1.0.0",
    "title": "Book Management"
  },
  "host": "localhost:3000",
  "basePath": "/api-docs/",
  "tags": [
    {
      "name": "book",
      "description": "Operations about book"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/addBook": {
      "post": {
        "tags": [
          "book"
        ],
        "summary": "Create Book",
        "description": "This can only be done by the logged in user.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/books": {
      "get": {
        "tags": [
          "book"
        ],
        "summary": "Get All Books",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/delete/id": {
      "delete": {
        "tags": [
          "book"
        ],
        "summary": "Delete book",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/addreviews/id": {
      "post": {
        "tags": [
          "book"
        ],
        "summary": "Add Reviews to a Book",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "",
            "schema": {
              "$ref": "#/definitions/Review"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/book/id": {
      "get": {
        "tags": [
          "book"
        ],
        "summary": "Get details of a Book",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Book": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "author": {
          "type": "string"
        },
        "isbn": {
          "type": "string"
        },
        "price": {
          "type": "string"
        }
      }
    },
    "Review": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "reviewername": {
          "type": "string"
        },
        "comments": {
          "type": "string"
        },
        "rating": {
          "type": "string"
        }
      }
    }
  }
}